requirements.txt
websocket-client
event_channel
deepl
speechrecognition
google-cloud-speech
google-cloud-texttospeech
pyaudio
pyttsx3
pytest
requests
six
pydub
gtts
simpleaudio
googletrans==3.1.0a0
lingua-language-detector
obs-websocket-py
openai
pyyaml
